chore(indexer/datasource): move pregod to kurora
