+++ service/indexer/internal/server/server.go
@@ -35,9 +35,9 @@ import (
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/blockscout"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/eip1577"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora"
+	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora/ethereum"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora/mirror"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/moralis"
-	eth_etl "github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/pregod_etl/ethereum"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/pregod_etl/lens"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/zksync"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource_asset"
@@ -164,7 +164,7 @@ func (s *Server) Initialize() (err error) {
 		return err
 	}
 
-	ethereumDatasource, err := eth_etl.New(context.Background(), s.config.Kurora.Endpoint)
+	ethereumDatasource, err := ethereum.New(context.Background(), s.config.Kurora.Endpoint)
 	if err != nil {
 		return err
 	}
